Fixed Viewport to use int instead of float. PiperOrigin-RevId: 140406823 Change-Id: Ifd54942d8771a9e380d3ddadc6fd6e054cc22a17

+++ include/fplbase/render_state.h
@@ -58,7 +58,7 @@ struct Viewport {
   mathfu::vec2i size;
 
   Viewport();
-  Viewport(float x, float y, float width, float height);
+  Viewport(int x, int y, int width, int height);
   Viewport(const mathfu::vec2i &pos, const mathfu::vec2i &size);
 };
